backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2